==> 1-merge-monthly-flight-data.ipynb <==
import os
import pandas as pd
import zipfile

def process_flight_data(folder_path, output_filename):
    master_df = pd.DataFrame()

    for filename in os.listdir(folder_path):
        if filename.endswith('.zip'):
            zip_path = os.path.join(folder_path, filename)
            with zipfile.ZipFile(zip_path, 'r') as z:
                for file_in_zip in z.namelist():
                    if file_in_zip.endswith('.csv'):
                        with z.open(file_in_zip) as my_file:
                            df = pd.read_csv(my_file)
                            master_df = pd.concat([master_df, df], ignore_index=True)
    
    master_df.to_csv(os.path.join(folder_path,output_filename))
    print(f"Data processed and saved to {output_filename}")



# os.chdir('C:\\Users\\smeshah\\Desktop\\flight-delay-prediction')

# Define the folder paths for each location based on your structure
locations = {
    "new-york": "./data/raw-monthly-flight-data/new-york",
    "alaska": "./data/raw-monthly-flight-data/alaska",
    "washington": "./data/raw-monthly-flight-data/washington"
}

# Process data for each location
for location, folder_path in locations.items():
    output_filename = f"{location.capitalize().replace('-', '_')}_Flight_Data_July_2023_June_2023.csv"
    process_flight_data(folder_path, output_filename)
