==> 2-weather-data.ipynb <==
import airportsdata
import numpy as np
import pandas as pd
from datetime import datetime, timedelta
import matplotlib.pyplot as plt
from meteostat import Point, Daily, Hourly, Stations
import openmeteo_requests
import requests_cache
from retry_requests import retry
from zoneinfo import ZoneInfo
import os

stations = Stations()
airports = airportsdata.load("IATA")

master_df = pd.read_csv("data/raw-monthly-flight-data/washington/Washington_Flight_Data_July_2023_June_2023.csv")

master_df.drop ('Unnamed: 0', axis = 1, inplace = True)
master_df.columns
master_df.info()

master_df['FL_DATE_ONLY'] = pd.to_datetime (master_df['FL_DATE']).dt.date

master_df.dropna (subset = ['DEP_TIME'], inplace = True)
master_df.dropna (subset = ['ARR_TIME'], inplace = True)

master_df['DEP_TIME'] = master_df['DEP_TIME'].astype (int)
master_df['ARR_TIME'] = master_df['ARR_TIME'].astype (int)

max_value = master_df['CRS_ARR_TIME'].max ()
master_df.loc[master_df['CRS_ARR_TIME'] == max_value, 'CRS_ARR_TIME'] = 0

max_value = master_df['ARR_TIME'].max ()
master_df.loc[master_df['ARR_TIME'] == max_value, 'ARR_TIME'] = 0

max_value = master_df['DEP_TIME'].max ()
master_df.loc[master_df['DEP_TIME'] == max_value, 'DEP_TIME'] = 0

master_df['CRS_DEP_TIME'] = master_df['CRS_DEP_TIME'].astype (str).str.zfill (4)
master_df['CRS_DEP_TIME'] = pd.to_datetime (master_df['CRS_DEP_TIME'], format = '%H%M').dt.time

master_df['CRS_ARR_TIME'] = master_df['CRS_ARR_TIME'].astype (str).str.zfill (4)
master_df['CRS_ARR_TIME'] = pd.to_datetime (master_df['CRS_ARR_TIME'], format = '%H%M').dt.time

master_df['DEP_TIME'] = master_df['DEP_TIME'].astype (str).str.zfill (4)
master_df['DEP_TIME'] = pd.to_datetime (master_df['DEP_TIME'], format = '%H%M').dt.time

master_df['ARR_TIME'] = master_df['ARR_TIME'].astype (str).str.zfill (4)
master_df['ARR_TIME'] = pd.to_datetime (master_df['ARR_TIME'], format = '%H%M').dt.time

display (master_df[['CRS_DEP_TIME', 'CRS_ARR_TIME', 'DEP_TIME', 'ARR_TIME']])

master_df['CRS_DEP_TIME_HOUR'] = master_df['CRS_DEP_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['CRS_ARR_TIME_HOUR'] = master_df['CRS_ARR_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['DEP_TIME_HOUR'] = master_df['DEP_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['ARR_TIME_HOUR'] = master_df['ARR_TIME'].apply (lambda x : round (x.hour + x.minute / 60))

master_df = master_df.reset_index()
master_df.head()

air_code = np.unique(master_df[['ORIGIN','DEST']].values)

def chunks(lst, n):
  chunk = list()
  i = 0
  while i<len(lst):
    chunk.append(lst[i:i+n])
    i = i+n
  return chunk

temp = list(air_code)
x = chunks(temp,12)

for i in x:
  print(i)

import time
weather_df = pd.DataFrame()
not_present = list()
count = 1
for i in x:
    print(i)
    weather_df = pd.DataFrame()
    for code in i:
        try:
            lat = airports[f'{code}']['lat']
            lon = airports[f'{code}']['lon']
        except:
            not_present.append(code)
            print("Not present -- ", not_present)
            continue
        cache_session = requests_cache.CachedSession('.cache', expire_after = -1)
        retry_session = retry(cache_session, retries = 5, backoff_factor = 0.2)
        openmeteo = openmeteo_requests.Client(session = retry_session)
        url = "https://archive-api.open-meteo.com/v1/archive"
        stations = stations.nearby(lat, lon)
        station = stations.fetch(1)
        st = station.reset_index()['id'][0]
        begin_date = datetime(2018, 1, 1)
        end_date = datetime(2023, 12, 31)
        getting_paras = list()
        getting_paras = {"temperature_2m","snowfall","relative_humidity_2m","precipitation","rain","cloud_cover","shortwave_radiation","wind_speed_100m","wind_direction_100m","wind_gusts_10m","snow_depth"}
        params = {
            "latitude": lat,
            "longitude": lon,
            "start_date": "2018-01-01",
            "end_date": "2023-07-31",
            "hourly": getting_paras
            }
        time.sleep(35)
        # Process hourly data. The order of variables needs to be the same as requested.
        responses = openmeteo.weather_api(url, params=params)
        # Process first location. Add a for-loop for multiple locations or weather models
        response = responses[0]
        print(f"Coordinates {response.Latitude()}°E {response.Longitude()}°N")
        print(f"Elevation {response.Elevation()} m asl")
        print(f"Timezone {response.Timezone()} {response.TimezoneAbbreviation()}")
        print(f"Timezone difference to GMT+0 {response.UtcOffsetSeconds()} s")
        hourly = response.Hourly()
        parameters_dataset = list()
        for i in range(11):
            parameters_dataset.append(hourly.Variables(i).ValuesAsNumpy())
        hourly_data = {"date": pd.date_range(
            start = pd.to_datetime(hourly.Time(), unit = "s"),
            end = pd.to_datetime(hourly.TimeEnd(), unit = "s"),
            freq = pd.Timedelta(seconds = hourly.Interval()),
            inclusive = "left"
        )}
        getting_paras = list(getting_paras)
        for i in range(11):
            hourly_data[getting_paras[i]] = list(parameters_dataset[i])
        hourly_dataframe = pd.DataFrame(data = hourly_data)
        print(f"location code {code} length {len(hourly_dataframe)}")
        hourly_dataframe['airport_code'] = code
        weather_df = pd.concat([hourly_dataframe,weather_df],ignore_index=True)
    weather_df.to_csv(f"data/weather-data-chunks/washington/new_weather_washington_{count}.csv")
    count+=1

print(not_present)
len(not_present)

Combining all datafiles

weather_df = pd.DataFrame()
directory = 'data/weather-data-chunks/washington/Washington_Flight_Data_July_2023_June_2023.csv'
for filename in os.listdir(directory):
    f = os.path.join(directory, filename)
    if os.path.isfile(f):
        print(f"reading {f}")
        curr = pd.read_csv(f)
        weather_df = pd.concat([curr,weather_df],ignore_index=True)
        print(f"{f} combined with main file")

weather_df.to_csv("data/weather-data/Washington_Weather_Data_July_2023_June_2023.csv")
