==> 3-merge-flight-and-weather-data.ipynb <==
import numpy as np
import pandas as pd
from zoneinfo import ZoneInfo
from datetime import datetime, timedelta
import meteostat
import airportsdata
airports = airportsdata.load ("IATA")

master_df = pd.read_csv ("data/raw-monthly-flight-data/washington/Washington_Flight_Data_July_2023_June_2023.csv")

new_weather_df = pd.read_csv("data/weather-data-chunks/washington/final_weather_df.csv")

master_df['FL_DATE_ONLY'] = pd.to_datetime (master_df['FL_DATE']).dt.date

master_df.dropna (subset = ['DEP_TIME'], inplace = True)
master_df.dropna (subset = ['ARR_TIME'], inplace = True)

master_df['DEP_TIME'] = master_df['DEP_TIME'].astype (int)
master_df['ARR_TIME'] = master_df['ARR_TIME'].astype (int)

max_value = master_df['CRS_ARR_TIME'].max ()
master_df.loc[master_df['CRS_ARR_TIME'] == max_value, 'CRS_ARR_TIME'] = 0

max_value = master_df['ARR_TIME'].max ()
master_df.loc[master_df['ARR_TIME'] == max_value, 'ARR_TIME'] = 0

max_value = master_df['DEP_TIME'].max ()
master_df.loc[master_df['DEP_TIME'] == max_value, 'DEP_TIME'] = 0

master_df['CRS_DEP_TIME'] = master_df['CRS_DEP_TIME'].astype (str).str.zfill (4)
master_df['CRS_DEP_TIME'] = pd.to_datetime (master_df['CRS_DEP_TIME'], format = '%H%M').dt.time

master_df['CRS_ARR_TIME'] = master_df['CRS_ARR_TIME'].astype (str).str.zfill (4)
master_df['CRS_ARR_TIME'] = pd.to_datetime (master_df['CRS_ARR_TIME'], format = '%H%M').dt.time

master_df['DEP_TIME'] = master_df['DEP_TIME'].astype (str).str.zfill (4)
master_df['DEP_TIME'] = pd.to_datetime (master_df['DEP_TIME'], format = '%H%M').dt.time

master_df['ARR_TIME'] = master_df['ARR_TIME'].astype (str).str.zfill (4)
master_df['ARR_TIME'] = pd.to_datetime (master_df['ARR_TIME'], format = '%H%M').dt.time

master_df['FL_DATE_ONLY'] = pd.to_datetime(master_df['FL_DATE_ONLY'])

master_df['DEP_TIME'] = pd.to_datetime (master_df['DEP_TIME'], format = '%H:%M:%S')

master_df['DEP_DATE_TIME'] = master_df['FL_DATE_ONLY'] + pd.to_timedelta (master_df['DEP_TIME'].dt.hour, unit = 'h') + pd.to_timedelta(master_df['DEP_TIME'].dt.minute, unit='m') + pd.to_timedelta(master_df['DEP_TIME'].dt.second, unit='s')

master_df['ARR_TIME'] = pd.to_datetime(master_df['ARR_TIME'], format='%H:%M:%S')

master_df['ARR_DATE_TIME'] = master_df['FL_DATE_ONLY'] + pd.to_timedelta(master_df['ARR_TIME'].dt.hour, unit='h') + pd.to_timedelta(master_df['ARR_TIME'].dt.minute, unit='m') + pd.to_timedelta(master_df['ARR_TIME'].dt.second, unit='s')

master_df['CRS_DEP_TIME'] = pd.to_datetime(master_df['CRS_DEP_TIME'], format='%H:%M:%S')

master_df['CRS_DEP_DATE_TIME'] = master_df['FL_DATE_ONLY'] + pd.to_timedelta(master_df['CRS_DEP_TIME'].dt.hour, unit='h') + pd.to_timedelta(master_df['CRS_DEP_TIME'].dt.minute, unit='m') + pd.to_timedelta(master_df['CRS_DEP_TIME'].dt.second, unit='s')

master_df['CRS_ARR_TIME'] = pd.to_datetime(master_df['CRS_ARR_TIME'], format='%H:%M:%S')

master_df['CRS_ARR_DATE_TIME'] = master_df['FL_DATE_ONLY'] + pd.to_timedelta(master_df['CRS_ARR_TIME'].dt.hour, unit='h') + pd.to_timedelta(master_df['CRS_ARR_TIME'].dt.minute, unit='m') + pd.to_timedelta(master_df['CRS_ARR_TIME'].dt.second, unit='s')

master_df['OP_UNIQUE_CARRIER'].unique()

convert_timezone = lambda row: row['CRS_DEP_DATE_TIME'].replace(tzinfo=ZoneInfo(airports[f"{row['ORIGIN']}"]['tz'])).astimezone(ZoneInfo('UTC'))

master_df['CRS_DEP_DATE_TIME'] = master_df.apply(convert_timezone, axis = 1)

convert_timezone = lambda row: row['DEP_DATE_TIME'].replace(tzinfo=ZoneInfo(airports[f"{row['ORIGIN']}"]['tz'])).astimezone(ZoneInfo('UTC'))

master_df['DEP_DATE_TIME'] = master_df.apply(convert_timezone, axis=1)

convert_timezone = lambda row: row['CRS_ARR_DATE_TIME'].replace(tzinfo=ZoneInfo(airports[f"{row['DEST']}"]['tz'])).astimezone(ZoneInfo('UTC'))

master_df['CRS_ARR_DATE_TIME'] = master_df.apply(convert_timezone, axis=1)

convert_timezone = lambda row: row['ARR_DATE_TIME'].replace(tzinfo=ZoneInfo(airports[f"{row['DEST']}"]['tz'])).astimezone(ZoneInfo('UTC'))

master_df['ARR_DATE_TIME'] = master_df.apply(convert_timezone, axis=1)

master_df['CRS_DEP_TIME_HOUR'] = master_df['CRS_DEP_DATE_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['CRS_ARR_TIME_HOUR'] = master_df['CRS_ARR_DATE_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['DEP_TIME_HOUR'] = master_df['DEP_DATE_TIME'].apply (lambda x : round (x.hour + x.minute / 60))
master_df['ARR_TIME_HOUR'] = master_df['ARR_DATE_TIME'].apply (lambda x : round (x.hour + x.minute / 60))

master_df = master_df.reset_index ()

max_value = master_df['CRS_DEP_TIME_HOUR'].max ()
master_df.loc[master_df['CRS_DEP_TIME_HOUR'] == max_value, 'CRS_DEP_TIME_HOUR'] = 0

max_value = master_df['CRS_ARR_TIME_HOUR'].max ()
master_df.loc[master_df['CRS_ARR_TIME_HOUR'] == max_value, 'CRS_ARR_TIME_HOUR'] = 0

max_value = master_df['DEP_TIME_HOUR'].max ()
master_df.loc[master_df['DEP_TIME_HOUR'] == max_value, 'DEP_TIME_HOUR'] = 0

max_value = master_df['ARR_TIME_HOUR'].max ()
master_df.loc[master_df['ARR_TIME_HOUR'] == max_value, 'ARR_TIME_HOUR'] = 0

new_weather_df[['date_only', 'time_only']] = new_weather_df['date'].str.split (pat = ' ', n = 1, expand = True)

new_weather_df['time_only'] = pd.to_datetime(new_weather_df['time_only'], format = '%H:%M:%S').dt.time

new_weather_df['hour_only'] = new_weather_df['time_only'].apply (lambda x : round (x.hour + x.minute / 60))

new_weather_df = new_weather_df.drop(columns=['Unnamed: 0.1'	,'Unnamed: 0'])

# new_weather_df = pd.read_csv("Data/washington_v2_new_weather_df_final_proposal.csv")
# master_df = pd.read_csv("Data/washington_v2_master_df_final_proposal.csv")

#master_df['FL_DATE_ONLY'] = pd.to_datetime (master_df['FL_DATE_ONLY'])
new_weather_df['date_only'] = pd.to_datetime (new_weather_df['date_only'])

new_weather_df = new_weather_df.rename(columns = {'airport_code':'airport'})

master_df['FL_DATE_ONLY'] = master_df['FL_DATE_ONLY'].astype('datetime64')

new_master_df1 = master_df.copy ()

merged_df = pd.merge (new_master_df1, new_weather_df, left_on = ['FL_DATE_ONLY', 'CRS_DEP_TIME_HOUR', 'ORIGIN'],
                      right_on = ['date_only', 'hour_only', 'airport'], how = 'inner')

merged_df = merged_df.drop (['date_only', 'hour_only', 'airport'], axis = 1)

new_columns = {
    #'Unnamed: 0': 'Unnamed: 0_DEP',
    'relative_humidity_2m' : 'relative_humidity_2m_DEP',
    'wind_gusts_10m' : 'wind_gusts_10m_DEP',
    'shortwave_radiation' : 'shortwave_radiation_DEP',
    'snowfall' : 'snowfall_DEP',
    'wind_direction_100m' : 'wind_direction_100m_DEP',
    'snow_depth' : 'snow_depth_DEP',
    'precipitation' : 'precipitation_DEP',
    'cloud_cover' : 'cloud_cover_DEP',
    'wind_speed_100m' : 'wind_speed_100m_DEP',
    'rain' : 'rain_DEP',
    'temperature_2m' : 'temperature_2m_DEP',
    'airport': 'airport_DEP',
    'time_only': 'time_only_DEP',
}

merged_df.rename (columns = new_columns, inplace = True)

# merged_df.to_csv ("Data/washington_v2_merged_df_DEP.csv")

new_master_df2 = merged_df.copy ()

final_merged_df = pd.merge (new_master_df2, new_weather_df, left_on = ['FL_DATE_ONLY', 'CRS_ARR_TIME_HOUR', 'DEST'],
                      right_on = ['date_only', 'hour_only', 'airport'], how = 'inner')

final_merged_df = final_merged_df.drop (['date_only', 'hour_only', 'airport'], axis = 1)

new_columns = {
    #'Unnamed: 0': 'Unnamed: 0_DEP',
    'relative_humidity_2m' : 'relative_humidity_2m_ARR',
    'wind_gusts_10m' : 'wind_gusts_10m_ARR',
    'shortwave_radiation' : 'shortwave_radiation_ARR',
    'snowfall' : 'snowfall_ARR',
    'wind_direction_100m' : 'wind_direction_100m_ARR',
    'snow_depth' : 'snow_depth_ARR',
    'precipitation' : 'precipitation_ARR',
    'cloud_cover' : 'cloud_cover_ARR',
    'wind_speed_100m' : 'wind_speed_100m_ARR',
    'rain' : 'rain_ARR',
    'temperature_2m' : 'temperature_2m_ARR',
    'airport': 'airport_ARR',
    'time_only': 'time_only_ARR',
}

final_merged_df.rename (columns = new_columns, inplace = True)

final_merged_df.to_csv ("data/merged-final-data/Washington_Merged_July_21_June_23.csv")

folder_path = 'data/merged-final-data/'  # Replace with your folder path
all_files = [os.path.join(folder_path, file) for file in os.listdir(folder_path) if file.endswith('.csv')]

all_dataframes = [] 
for file in all_files:
    df = pd.read_csv(file)
    all_dataframes.append(df)

concatenated_df = pd.concat(all_dataframes, ignore_index=True)

concatenated_df.to_csv('data/merged-final-data/Final_Merged_July_21_June_23.csv',index = False)
